+++ NEWS.md
@@ -1,6 +1,7 @@
 # huxtable 0.3.0.9000 (development version on github)
 
 * New convenience functions `insert_row` and `insert_column`.
+* `latex_float` property allows you to change positioning in LaTeX.
 
 # huxtable 0.3.0